change bitrate option, ffmpeg downloading, switch to eyed3
new option to change bitrate to 96, 128, 192, or 320kbps. useful for users with limited storage, while keeping the audiophiles happy.
switch to downloading with ffmpeg, and changing ogg to mp3. (eyed3 can only apply metadata such as album art to an mp3)
addition of album art with eyed3 (finally), as well as fully dumping mutagen for eyed3 metadata applying.