remove known issues (all resolved)
i didn't realize but by making ffmpeg handle the downloading, it removes the issue of the connection reset error and the aiohttp server disconnected error which were occurring, as ffmpeg now handles all of those things, not me. this makes natsuka essentially bug free at the moment as far as i know, unless i find any bugs in the future.