auto download to downloads folder on android
i made this change as more than likely if you're using this on termux, you're using it to play within a music player anyway, and unless you have root, the music if downloaded in the termux environment will be inaccessible to file explorers and music players.
in the event you do want to actually play this music in termux (with a player like mpv or vlc), you can do
```
termux-setup-storage
```

and allow storage permissions to termux. this will allow you to access the system folders in termux and be able to play from `storage/downloads/Music`.